+++ vite.config.ts
@@ -4,6 +4,7 @@ import { defineConfig } from 'vite';
 export default defineConfig({
 	plugins: [sveltekit()],
 	server: {
-		port: 3000
+		port: 3000,
+		allowedHosts: ['ignignokt', 'ignignokt.home.youngsecurity.net']
 	}
 });